improve error handling

+++ src/deleteFaunaDocument.ts
@@ -6,6 +6,7 @@ const deleteFaunaDocument = async (clientOrToken: Client | string, classes: stri
   let client: Client
   if (typeof clientOrToken === 'string') client = faunaClient(clientOrToken)
   else client = clientOrToken
+  if (client === undefined) throw new Error('Fauna client is undefined')
   return client
     .query(faunaQuery.Delete(faunaQuery.Ref(`classes/${classes}/${id}`)))
     .then((response: any) => {
